Changed the background, formatting
Almost done with the website design!! Just have to do a few more changes and I should be good to go! I organized the website a bit more and tried to make the tag pages look a bit nicer. I still need to write out a few blog posts, so right now it's just filler content, but I will do so as soon as I feel a bit more satisfied with what I have!!