Blog Post: Typora Marigins
Made a short post on how to change the margins in Typora when printing. Need to do something about the <code> tags formatting later, but for now it's fine. Made a couple minor changes such as fixing a few links that I didn't realize were wrong as well as dates (was using 2020 instead of 2021 xD)